Transform system has been revamped and dx12 has been updated
-The transform array starts off with the max dynamic object count
-Dynamic transforms are added to those initial elements, and static ones
are pushed back
-Every transform is held on the same buffer but dx12 only updates the
part where the dynamic transform resides
-Performance seems to improve by quite a bit. Vulkan will receive the
update as well

+++ src/BlitzenDX12/dx12Data.h
@@ -125,7 +125,7 @@ namespace BlitzenDX12
 
         DX12WRAPPER<ID3D12Resource> staging{ nullptr };
         void* pData{ nullptr };
-        uint32_t dataCopySize{ 0 };
+        size_t dataCopySize{ 0 };
     };
